Adds skeleton for EOL module
