+++ petal/mining/modules/__init__.py
@@ -1,2 +1,3 @@
 from .wikipedia import WikipediaModule
 from .backbone  import BackboneModule
+from .eol       import EOLModule
